Move hash_from_code to theano.utils and remove theano.gof.utils.hash_from_file
