+++ theano/sparse/utils.py
@@ -1,4 +1,4 @@
-from theano.gof.utils import hash_from_code
+from theano.utils import hash_from_code
 
 
 def hash_from_sparse(data):
